各チームの行を列挙するViewに分割
