+++ MolkkyScoreBoard/MolkkyScoreBoard/UI/MolkkyPlay/Views/Lower/TeamScoresView.swift
@@ -1,0 +1,40 @@
+//
+//  TeamScoresView.swift
+//  MolkkyScoreBoard
+//
+//  Created by ta9yamakawa on 2023/03/03.
+//
+
+import SwiftUI
+import ComposableArchitecture
+
+/// 各チームのスコアView
+struct TeamScoresView: View {
+    /// View Store
+    let viewStore: ViewStore<MolkkyPlayFeature.State,
+                             MolkkyPlayFeature.Action>
+
+    var body: some View {
+        VStack(spacing: 0) {
+            TeamScoresTitleView()
+            TeamScoresRowsView(viewStore: viewStore)
+        }
+    }
+}
+
+// MARK: Previews
+struct TeamScoresView_Previews: PreviewProvider {
+    static var previews: some View {
+        let teams = [Team(id: 1,
+                          members: [TeamMember(name: "hoge")],
+                          order: 0, isDisqualified: true),
+                     Team(id: 2,
+                          members: [TeamMember(name: "huga")],
+                          order: 1)]
+        let state = MolkkyPlayFeature.State(teams: teams,
+                                            isLatterHalf: false)
+        let viewStore = ViewStore(StoreOf<MolkkyPlayFeature>(initialState: state,
+                                                             reducer: MolkkyPlayFeature()))
+        TeamScoresView(viewStore: viewStore)
+    }
+}
